# docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  handler.SignupRequest:
    properties:
      confirm_password:
        minLength: 8
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      user_currency:
        maxLength: 3
        minLength: 3
        type: string
    required:
    - confirm_password
    - email
    - first_name
    - last_name
    - password
    - user_currency
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
      user_currency:
        type: string
      user_id:
        type: string
      user_role:
        description: user or admin
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a sample server Instashop server.
  title: Instashop Swagger API
  version: "1.0"
paths:
...
  /api/v1/user/login:
    post:
      description: Authenticates a user using their email and password
      parameters:
      - description: Login Request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful authentication
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  ' expires':
                    type: string
                  ' issued':
                    type: string
                  ' issuer':
                    type: string
                  ' user_id':
                    type: string
                  token:
                    type: string
                type: object
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  ' message':
                    type: string
                  error:
                    type: boolean
                type: object
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  ' message':
                    type: string
                  error:
                    type: boolean
                type: object
            type: object
      summary: Authenticate a user
      tags:
      - Authentication
...
  /api/v1/user/signup:
    post:
      description: Registers a user with the provided details
      parameters:
      - description: Signup Request
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/handler.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User successfully registered
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  ' order':
                    $ref: '#/definitions/model.User'
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  ' message':
                    type: string
                  error:
                    type: boolean
                type: object
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  ' message':
                    type: string
                  error:
                    type: boolean
                type: object
            type: object
      summary: Register a new user
      tags:
      - Authentication
swagger: "2.0"